Refactor data
