+++ docs/source/api_reference/model.rst
@@ -40,6 +40,15 @@ Currency
    :members:
    :member-order: bysource
 
+Data
+----
+
+.. automodule:: nautilus_trader.model.data
+   :show-inheritance:
+   :inherited-members:
+   :members:
+   :member-order: bysource
+
 Enums
 -----
