+++ nautilus_trader/trading/strategy.pyx
@@ -42,12 +42,9 @@ from nautilus_trader.common.logging cimport LogColor
 from nautilus_trader.common.logging cimport Logger
 from nautilus_trader.common.logging cimport RECV
 from nautilus_trader.common.logging cimport REQ
-from nautilus_trader.common.logging cimport RES
 from nautilus_trader.common.logging cimport SENT
 from nautilus_trader.core.constants cimport *  # str constants only
 from nautilus_trader.core.correctness cimport Condition
-from nautilus_trader.data.base cimport Data
-from nautilus_trader.data.base cimport DataType
 from nautilus_trader.data.engine cimport DataEngine
 from nautilus_trader.data.messages cimport DataRequest
 from nautilus_trader.data.messages cimport Subscribe
@@ -60,6 +57,8 @@ from nautilus_trader.model.commands cimport AmendOrder
 from nautilus_trader.model.commands cimport CancelOrder
 from nautilus_trader.model.commands cimport SubmitBracketOrder
 from nautilus_trader.model.commands cimport SubmitOrder
+from nautilus_trader.model.data cimport DataType
+from nautilus_trader.model.data cimport GenericData
 from nautilus_trader.model.events cimport Event
 from nautilus_trader.model.events cimport OrderCancelReject
 from nautilus_trader.model.events cimport OrderDenied
@@ -389,13 +388,13 @@ cdef class TradingStrategy(Component):
         """
         pass  # Optionally override in subclass
 
-    cpdef void on_data(self, Data data) except *:
+    cpdef void on_data(self, GenericData data) except *:
         """
         Actions to be performed when the strategy is running and receives a data object.
 
         Parameters
         ----------
-        data : Data
+        data : GenericData
             The data received.
 
         Warnings
@@ -1769,7 +1768,7 @@ cdef class TradingStrategy(Component):
         for i in range(length):
             self.handle_bar(bar_type, bars[i], is_historical=True)
 
-    cpdef void handle_data(self, Data data) except *:
+    cpdef void handle_data(self, GenericData data) except *:
         """
         Handle the given data.
 
@@ -1777,7 +1776,7 @@ cdef class TradingStrategy(Component):
 
         Parameters
         ----------
-        data : Data
+        data : GenericData
             The received data.
 
         Warnings
