+++ tests/unit_tests/model/test_model_tick.py
@@ -63,15 +63,9 @@ class QuoteTickTests(unittest.TestCase):
 
         self.assertTrue(tick1 == tick1)
         self.assertTrue(tick1 != tick2)
-        self.assertTrue(tick1 <= tick1)
-        self.assertTrue(tick1 <= tick2)
-        self.assertTrue(tick1 < tick2)
-        self.assertTrue(tick3 > tick2)
-        self.assertTrue(tick3 >= tick2)
-        self.assertTrue(tick3 >= tick3)
-        self.assertEqual([tick1, tick2, tick3], sorted([tick2, tick3, tick1]))
+        self.assertEqual([tick1, tick2, tick3], sorted([tick2, tick3, tick1], key=lambda x: x.timestamp))
 
-    def test_tick_hash_str_and_repr(self):
+    def test_tick_str_and_repr(self):
         # Arrange
         tick = QuoteTick(
             AUDUSD_SIM.security,
@@ -87,7 +81,6 @@ class QuoteTickTests(unittest.TestCase):
         result1 = repr(tick)
 
         # Assert
-        self.assertEqual(int, type(hash(tick)))
         self.assertEqual("AUD/USD.SIM,1.00000,1.00001,1,1,1970-01-01T00:00:00.000Z", result0)
         self.assertEqual("QuoteTick(AUD/USD.SIM,1.00000,1.00001,1,1,1970-01-01T00:00:00.000Z)", result1)
 
@@ -242,14 +235,7 @@ class TradeTickTests(unittest.TestCase):
         )
 
         self.assertTrue(tick1 == tick1)
-        self.assertTrue(tick1 != tick2)
-        self.assertTrue(tick1 <= tick1)
-        self.assertTrue(tick1 <= tick2)
-        self.assertTrue(tick1 < tick2)
-        self.assertTrue(tick3 > tick2)
-        self.assertTrue(tick3 >= tick2)
-        self.assertTrue(tick3 >= tick3)
-        self.assertEqual([tick1, tick2, tick3], sorted([tick2, tick3, tick1]))
+        self.assertEqual([tick1, tick2, tick3], sorted([tick2, tick3, tick1], key=lambda x: x.timestamp))
 
     def test_str_and_repr(self):
         # Arrange
