+++ tests/unit_tests/trading/test_trading_strategy.py
@@ -30,14 +30,14 @@ from nautilus_trader.common.enums import ComponentState
 from nautilus_trader.common.logging import TestLogger
 from nautilus_trader.common.uuid import UUIDFactory
 from nautilus_trader.core.fsm import InvalidStateTrigger
-from nautilus_trader.data.base import Data
-from nautilus_trader.data.base import DataType
 from nautilus_trader.data.engine import DataEngine
 from nautilus_trader.execution.database import BypassExecutionDatabase
 from nautilus_trader.execution.engine import ExecutionEngine
 from nautilus_trader.indicators.average.ema import ExponentialMovingAverage
 from nautilus_trader.model.bar import Bar
 from nautilus_trader.model.currencies import USD
+from nautilus_trader.model.data import DataType
+from nautilus_trader.model.data import GenericData
 from nautilus_trader.model.enums import AssetClass
 from nautilus_trader.model.enums import AssetType
 from nautilus_trader.model.enums import OMSType
@@ -307,7 +307,7 @@ class TradingStrategyTests(unittest.TestCase):
         strategy = TradingStrategy("000")
 
         # Act
-        strategy.on_data(Data(DataType(str), "DATA"))
+        strategy.on_data(GenericData(DataType(str), "DATA", UNIX_EPOCH))
 
         # Assert
         self.assertTrue(True)  # Exception not raised
@@ -666,7 +666,7 @@ class TradingStrategyTests(unittest.TestCase):
 
         # Act
         # Assert
-        self.assertRaises(RuntimeError, strategy.handle_data, Data(DataType(str), "SOME_DATA"))
+        self.assertRaises(RuntimeError, strategy.handle_data, GenericData(DataType(str), "SOME_DATA", UNIX_EPOCH))
 
     def test_handle_event_when_user_code_raises_exception_logs_and_reraises(self):
         # Arrange
@@ -1300,7 +1300,7 @@ class TradingStrategyTests(unittest.TestCase):
             self.logger,
         )
 
-        data = Data(DataType(str), "SOME_DATA")
+        data = GenericData(DataType(str), "SOME_DATA", UNIX_EPOCH)
 
         # Act
         strategy.handle_data(data)
@@ -1319,7 +1319,7 @@ class TradingStrategyTests(unittest.TestCase):
 
         strategy.start()
 
-        data = Data(DataType(str), "SOME_DATA")
+        data = GenericData(DataType(str), "SOME_DATA", UNIX_EPOCH)
 
         # Act
         strategy.handle_data(data)
